demo_pagination/requirements.txt
Flask>=2.1
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.0
marshmallow>=3.0
marshmallow-sqlalchemy>=0.26
pytest>=7.0
pytest-flask>=1.2